Removed rho check, moved tests to src/tests.rs
